--no-build/rds_stopper.py

+++ deploy-prod.sh
@@ -1,4 +1,5 @@
 NOTESTS=0
+NOBUILD=0
 NOCACHE=0
 TAG=latest
 
@@ -6,6 +7,10 @@ while [[ $# -gt 0 ]]; do
     case $1 in
         --no-tests)
             NOTESTS=1
+            shift
+            ;;
+        --no-build)
+            NOBUILD=1
             shift
             ;;
         --no-cache)
@@ -34,40 +39,42 @@ fi
 DOCKER_SERVER=${AWS_ACCOUNT_ID}.dkr.ecr.${AWS_REGION}.amazonaws.com
 DOCKER_IMAGE=${DOCKER_SERVER}/${AWS_ECR_IMAGE_NAME}:${TAG}
 
-aws ecr get-login-password | docker login --username AWS --password-stdin ${DOCKER_SERVER}
-if [ $? -ne 0 ]; then
-    exit 1
-fi
+if [ $NOBUILD -eq 0 ]; then
+    aws ecr get-login-password | docker login --username AWS --password-stdin ${DOCKER_SERVER}
+    if [ $? -ne 0 ]; then
+        exit 1
+    fi
 
-if [ $NOTESTS -eq 0 ]; then
-    pytest
+    if [ $NOTESTS -eq 0 ]; then
+        pytest
+
+        if [ $? -ne 0 ]; then
+            exit 1
+        fi
+    else
+        echo "Skipping tests..."
+    fi
+
+    zappa save-python-settings-file
+    if [ $NOCACHE -eq 1 ]; then
+        docker build --no-cache -t ${DOCKER_IMAGE} .
+    else
+        docker build -t ${DOCKER_IMAGE} .
+    fi
+    rm zappa_settings.py
 
     if [ $? -ne 0 ]; then
         exit 1
     fi
-else
-    echo "Skipping tests..."
+
+    docker push ${DOCKER_IMAGE}
+    if [ $? -ne 0 ]; then
+        exit 1
+    fi
 fi
 
-zappa save-python-settings-file
-if [ $NOCACHE -eq 1 ]; then
-    docker build -t --no-cache ${DOCKER_IMAGE} .
-else
-    docker build -t ${DOCKER_IMAGE} .
-fi
-rm zappa_settings.py
-
-if [ $? -ne 0 ]; then
-    exit 1
-fi
-
-docker push ${DOCKER_IMAGE}
-if [ $? -ne 0 ]; then
-    exit 1
-fi
-
-if [ $(zappa status app -j) == *"Error: No Lambda aws-portal-app detected"* ]
-then
+zappa status app &> /dev/null
+if [ $? -eq 1 ]; then
     zappa deploy app -d ${DOCKER_IMAGE}
 else
     zappa update app -d ${DOCKER_IMAGE}
